Fix data type conversion for XGBoost

+++ backend/main.py
@@ -64,22 +64,22 @@ class SimulationResult(BaseModel):
 def prepare_features_for_prediction(home_cf, away_cf, home_xgf, away_xgf, home_hdcf, away_hdcf, home_sf, away_sf, home_gf, away_gf):
     """Prepare features for model prediction"""
     features = pd.DataFrame({
-        'home_cf_pct': [home_cf],
-        'away_cf_pct': [away_cf],
-        'cf_diff': [home_cf - away_cf],
-        'home_xgf_pct': [home_xgf],
-        'away_xgf_pct': [away_xgf],
-        'xgf_diff': [home_xgf - away_xgf],
-        'home_hdcf_pct': [home_hdcf],
-        'away_hdcf_pct': [away_hdcf],
-        'hdcf_diff': [home_hdcf - away_hdcf],
-        'home_sf_pct': [home_sf],
-        'away_sf_pct': [away_sf],
-        'sf_diff': [home_sf - away_sf],
-        'home_gf_pct': [home_gf],
-        'away_gf_pct': [away_gf],
-        'gf_diff': [home_gf - away_gf],
-        'home_advantage': [1]
+        'home_cf_pct': [float(home_cf)],
+        'away_cf_pct': [float(away_cf)],
+        'cf_diff': [float(home_cf) - float(away_cf)],
+        'home_xgf_pct': [float(home_xgf)],
+        'away_xgf_pct': [float(away_xgf)],
+        'xgf_diff': [float(home_xgf) - float(away_xgf)],
+        'home_hdcf_pct': [float(home_hdcf)],
+        'away_hdcf_pct': [float(away_hdcf)],
+        'hdcf_diff': [float(home_hdcf) - float(away_hdcf)],
+        'home_sf_pct': [float(home_sf)],
+        'away_sf_pct': [float(away_sf)],
+        'sf_diff': [float(home_sf) - float(away_sf)],
+        'home_gf_pct': [float(home_gf)],
+        'away_gf_pct': [float(away_gf)],
+        'gf_diff': [float(home_gf) - float(away_gf)],
+        'home_advantage': [1.0]
     })
     return features
 
@@ -115,16 +115,16 @@ async def simulate(request: SimulationRequest):
         raise HTTPException(status_code=400, detail="No test data available for selected seasons")
     
     X_test = prepare_features_for_prediction(
-        test_df['home_cf_pct'].values,
-        test_df['away_cf_pct'].values,
-        test_df['home_xgf_pct'].values,
-        test_df['away_xgf_pct'].values,
-        test_df['home_hdcf_pct'].values,
-        test_df['away_hdcf_pct'].values,
-        test_df['home_sf_pct'].values,
-        test_df['away_sf_pct'].values,
-        test_df['home_gf_pct'].values,
-        test_df['away_gf_pct'].values
+        test_df['home_cf_pct'].astype(float).values,
+        test_df['away_cf_pct'].astype(float).values,
+        test_df['home_xgf_pct'].astype(float).values,
+        test_df['away_xgf_pct'].astype(float).values,
+        test_df['home_hdcf_pct'].astype(float).values,
+        test_df['away_hdcf_pct'].astype(float).values,
+        test_df['home_sf_pct'].astype(float).values,
+        test_df['away_sf_pct'].astype(float).values,
+        test_df['home_gf_pct'].astype(float).values,
+        test_df['away_gf_pct'].astype(float).values
     )
     
     probabilities = model.predict_proba(X_test)[:, 1]
